backend/requirements.txt
streamlit>=1.29.0
boto3>=1.34.0
pandas>=2.1.0
python-dotenv>=1.0.0
strands-agents>=0.1.0
numpy>=1.26.0
numexpr>=2.14.0
requests>=2.31.0

# MCP and AgentCore Gateway dependencies
mcp>=1.0.0

# AgentCore Runtime dependencies
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0
